finish database

+++ database/migrations/2019_12_29_084821_investors.php
@@ -16,9 +16,9 @@ class Investors extends Migration
         Schema::create('investors', function (Blueprint $table) {
             $table->increments('idIn');
             $table->string('name', 20);
-            $table->tinyInteger('phone');
+            $table->bigInteger('phone');
             $table->string('email', 30);
-            $table->decimal('price', 15, 3);
+            $table->float('price', 12, 2);
             $table->string('date', 10);
         });
     }
